backend/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis和Celery
redis==5.0.1
celery==5.3.4

# 交易所API
ccxt==4.1.40

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 工具
python-dateutil==2.8.2
pytz==2023.3